+++ api/src/main/java/javax/mvc/security/CsrfValid.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package javax.mvc.annotation;
+package javax.mvc.security;
 
 import javax.ws.rs.NameBinding;
 import java.lang.annotation.Documented;
